requirements.txt
imutils
pillow
Numpy
opencv-python
opencv-python-headless 
matplotlib
tqdm
labelImg
ultralytics